Remove typedefs and rename structures
Going off of https://www.kernel.org/doc/Documentation/CodingStyle,
chapter 5, one should not use a typedef on structures to 'hide' the
fact that they are structures. Hiding them prevents a) easily telling
at a glance what is and is not a type, and b) prevents editors from
highlighting the type name, since without the hint of 'struct', they
have no idea what is simply a new variable or a new type. Also, it is
customary to name new user-created types with '_t' at the end
(ala my_new_type_t), so that was done as well.