Move files for organization and add Makefile for easier building
It's atypical to just leave source and headers files in the project
root directory, so moved them into typical places. Added the Makefile
so that one doesn't have to enter a long compliation command.

On the note of the Makefile, during testing, I found that compiling
with optimizations (passing "-O2" to gcc) caused the compiler to emit
optimizations that caused a SIGSEGV. On closer inspection, it appears
that the compiler is modifying the array 'str' in showTable(), causing
a strcat() to fail later in sequence, causing said SIGSEGV. This should
be looked at.