Fix input handling to be less derpy
scanf() isn't a good function. It's very brittle and will easily
break from input that doesn't match its pattern string. It's better
to replace it with a process of reading input with a simple read
function (fgets() here), and convert with a conversion function
(strtod() here). This allows for several things :

1. Allowing for input to not match the pattern
(Good for allowing commands like 'quit')
2. Catching the fact that there was no actual input
3. Catching input that is malformed, and not allow it to pass

The new implementation allows for all of these things.