Fix all memory leaks
Ultimately, the memory leaks were caused by several things -
1. numToStr() was allocating memory for strings that wasn't
being free()'d. Eventually, we should thing about handling
strings a bit better (for example, in Lua, there is only
one copy of a string, reducing memory usage), but for now
changing the variable to 'static char[50]' means that the
burden of managing the memory falls from us to the compiler.

2. For some reason, there wasn't a call to delTable() to get
rid of table1 at the end of main(). This has been rectified.

3. delTable() was using it's own freeing methods, which /may/
have been causing problems. Regardless, changed it to use
rmvFromTable() for better consistancy.

4. Having placed a call to delTable(), I found that the
algorithm was skipping over the first element of the table,
causing an element to leak anyways. Currently, there is a
special-case codepath for the first element, but this temporary.
The better solution would be to fix the algorithm to handle
deletions of the first element, but I was more concerned to
fix the leaks. Marked with a TODO.